Fix typed bindings
The various functions that used typed bindings were failing to bind
the entityKey and parameter fields because the JsonProperty attributes
were only visible to Newtonsoft not the ASP.NET Web runtime.

To address this the fields of LevelCompleteRequest/EntityRequest<T>
have been renamed to match the actual field names in the JSON request
and the JsonProperty attributes have been removed.

It appears that the top level object has to have get; set; properties
in order to work as a typed parameter, but nested types can just have
public fields.